UseNLog - Replaced dynamic assembly loading with registration methods
