+++ examples/ASP.NET 4.6.1/Visual Studio 2017/ASP.NET 4.6.1 - VS2017/Scripts/jquery-3.4.1.intellisense.js
