+++ examples/ASP.NET 4.6.1/Visual Studio 2017/ASP.NET 4.6.1 - VS2017/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="Antlr" version="3.5.0.2" targetFramework="net461" />
   <package id="bootstrap" version="4.3.1" targetFramework="net461" />
-  <package id="jQuery" version="3.3.1" targetFramework="net461" />
+  <package id="jQuery" version="3.4.1" targetFramework="net461" />
   <package id="jQuery.Validation" version="1.19.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.9" targetFramework="net461" />
   <package id="Microsoft.AspNet.Razor" version="3.2.9" targetFramework="net461" />
@@ -12,7 +12,7 @@
   <package id="Microsoft.jQuery.Unobtrusive.Validation" version="3.1.2" targetFramework="net461" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net461" />
   <package id="Modernizr" version="2.8.3" targetFramework="net461" />
-  <package id="NLog" version="5.1.3" targetFramework="net461" />
+  <package id="NLog" version="5.2.0" targetFramework="net461" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.4.1" targetFramework="net461" />
   <package id="WebGrease" version="1.6.0" targetFramework="net461" />
 </packages>
