+++ src/NLog.Web/Internal/AssemblyExtensionTypes.tt
@@ -1,0 +1,123 @@
+<#@ template debug="false" hostspecific="false" language="C#" #>
+<#@ assembly name="System.Core" #>
+<#@ assembly name="$(TargetDir)NLog.dll" #>
+<#@ assembly name="$(TargetDir)NLog.Web.dll" #>
+<#@ import namespace="System.Collections.Generic" #>
+<#@ import namespace="System.Linq" #>
+<#@ import namespace="System.Reflection" #>
+<#@ import namespace="System.Text" #>
+<#@ output extension=".cs" #>
+<# 
+    //T4 templates are built in Visual Studio. See https://msdn.microsoft.com/en-us/library/bb126445.aspx
+#>// 
+// Copyright (c) 2004-2021 Jaroslaw Kowalski <jaak@jkowalski.net>, Kim Christensen, Julian Verdurmen
+// 
+// All rights reserved.
+// 
+// Redistribution and use in source and binary forms, with or without 
+// modification, are permitted provided that the following conditions 
+// are met:
+// 
+// * Redistributions of source code must retain the above copyright notice, 
+//   this list of conditions and the following disclaimer. 
+// 
+// * Redistributions in binary form must reproduce the above copyright notice,
+//   this list of conditions and the following disclaimer in the documentation
+//   and/or other materials provided with the distribution. 
+// 
+// * Neither the name of Jaroslaw Kowalski nor the names of its 
+//   contributors may be used to endorse or promote products derived from this
+//   software without specific prior written permission. 
+// 
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE 
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE 
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR 
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN 
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF 
+// THE POSSIBILITY OF SUCH DAMAGE.
+// 
+
+namespace NLog.Web.Internal
+{
+    /// <summary>
+    /// Provides logging interface and utility functions.
+    /// </summary>
+    internal static class AssemblyExtensionTypes
+    {
+        public static void RegisterTypes(this NLog.Config.ISetupExtensionsBuilder setupBuilder)
+        {
+            #pragma warning disable CS0618 // Type or member is obsolete
+<#
+    var types  = typeof(NLog.Web.Targets.AspNetTraceTarget).Assembly.GetTypes().OrderBy(t => t.ToString());
+
+    var net35 = new string[]
+    {
+        "NLog.Web.LayoutRenderers.AspNetUserClaimLayoutRenderer",
+    };
+
+    foreach(var type in types)
+    {
+        if (type.IsAbstract || type.IsPrimitive || !type.IsPublic)
+            continue;
+
+        if (net35.Contains(type.ToString()))
+            continue;
+
+        if (typeof(NLog.Targets.Target).IsAssignableFrom(type))
+        {
+            var targetAttributes = type.GetCustomAttributes<NLog.Targets.TargetAttribute>(false);
+            foreach (var targetAlias in targetAttributes)
+            {
+                var targetAliasName = targetAlias.Name;
+#>
+            setupBuilder.RegisterTarget<<#= type #>>("<#= targetAliasName #>");
+<#
+            }
+        }
+        else if (typeof(NLog.Layouts.Layout).IsAssignableFrom(type))
+        {
+            var layoutAttributes = type.GetCustomAttributes<NLog.Layouts.LayoutAttribute>(false);
+            foreach (var layoutAlias in layoutAttributes)
+            {
+                var layoutAliasName = layoutAlias.Name;
+#>
+            setupBuilder.RegisterLayout<<#= type #>>("<#= layoutAliasName #>");
+<#
+            }
+        }
+        else if (typeof(NLog.LayoutRenderers.LayoutRenderer).IsAssignableFrom(type))
+        {
+            var layoutAttributes = type.GetCustomAttributes<NLog.LayoutRenderers.LayoutRendererAttribute>(false);
+            foreach (var layoutAlias in layoutAttributes)
+            {
+                var layoutAliasName = layoutAlias.Name;
+#>
+            setupBuilder.RegisterLayoutRenderer<<#= type #>>("<#= layoutAliasName #>");
+<#
+            }
+
+            var ambientAttributes = type.GetCustomAttributes<NLog.LayoutRenderers.AmbientPropertyAttribute>(false);
+            foreach (var layoutAlias in ambientAttributes)
+            {
+                var layoutAliasName = layoutAlias.Name;
+#>
+            setupBuilder.RegisterAmbientRenderer<<#= type #>>("<#= layoutAliasName #>");
+<#
+            }
+        }
+    }
+#>
+
+#if !NET35
+            setupBuilder.RegisterLayoutRenderer<NLog.Web.LayoutRenderers.AspNetUserClaimLayoutRenderer>("aspnet-user-claim");
+#endif
+
+            #pragma warning restore CS0618 // Type or member is obsolete
+        }
+    }
+}
